Rename MapFloor to Floor [MA-119]
